api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tensorflow-cpu>=2.14.0  # Use CPU version to save money on cloud hosting!
pandas>=2.0.0
scikit-learn>=1.3.0